% Auto-generated: do not edit by hand
\name{bananaNode}

\alias{bananaNode}

\title{BananaNode component}

\description{
ddd
}

\usage{
bananaNode(id=NULL, data=NULL, selected=NULL)
}

\arguments{
\item{id}{Character. ddd}

\item{data}{Lists containing elements 'label', 'depends_on', 'node_id', 'onchange', 'ondelete', 'dependsonoptions'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional)
  - depends_on (unnamed list; optional)
  - node_id (character; optional)
  - onchange (optional)
  - ondelete (optional)
  - dependsonoptions (logical | numeric | character | named list | unnamed list; optional). ddd}

\item{selected}{Logical. ddd}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

