% Auto-generated: do not edit by hand
\name{resizableNode}

\alias{resizableNode}

\title{ResizableNode component}

\description{
ddd
}

\usage{
resizableNode(data=NULL, selected=NULL)
}

\arguments{
\item{data}{Lists containing elements 'label'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional). ddd}

\item{selected}{Logical. ddd}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

